Totally changed the way that VALUEs (Lisp objects) are defined; instead of wasting memory storing type information in cons cells and storing integers as allocated objects, we now do some bit-twiddling to store a subset of the type information in the two or three lowest bits of the VALUE.
This means that cons cells only take 2 words (8 bytes) instead of 3,
and integers can be stored in the VALUE itself (but now we only have
29-bit signed integers, not 32). Also we have to enforce strict
alignment rules: cons cells to at least eight bytes, _everything_ else
to at least four bytes.

It's basically trading memory use in the data section, for some code
bloat (the extra instructions to do the bit twiddling). It's not too
bad though, all objects but cons cells and numbers can be accessed from
a VALUE as-is, once their type has been established (which does take a
few more instructions).

I've also taken this opportunity to clean up some of the other Lisp
object handling code. Also, the plan is to remove position objects
completely and store them as a cons cell containing two integers.

To overcome the loss of the three bits in integers, I've defined a new
pseudo-type called LONG_INT, which is a cons cell containing the low 24
bits in its car, and the high bits in the cdr. This allows large values
like time stamps to be stored without dropping bits.