Bugfix in byte compiler of lost docstring.
Practical effect: Previously, if you byte compile, then the doc string
is lost in sawfish-config. This is fixed.

Details: Loss of docstring happened only if (1) byte compiler is not given
--write-docs (2) defvar is invoked within macro definition. Plain defvar
was not affected by this bug.
It was because in 'trans-defvar' in lisp/rep/vm/compiler/rep.jl, the
docstring is stored in a wrong property 'variable-documentation'. It
is corrected to 'documentation'.

Symbol 'variable-documentation' is not used elsewhere, including Sawfish
and emacs' sawfish-mode.